build appstore packages with symbols/add targets to merge into universal build
while there is no builtin support to create a fat binary during the
actual build phase, we have a post-processing script that can merge two
separately built architectures into an universal binary one.
Thus enhance the makefile to support a more convenient way to build the
two archs without the need to manually move the package out of the way
to build the other arch.

1) assemble one architecture's package via
   "make mac-app-store-package.prepare"
2) make clean and change targets in autogen.input to build for the other
   architecture
3) build the other architecture and combine into universal binary and
   also package into the final deliverable via
   make mac-app-store-package.universal

(of course you can also use the .prepare target separately also in the
second build, and just mac-app-store-package will build the deliverable
for the current architecture like it did before)

Change-Id: I02c14e9b8c4f6e39d0d50cd3b4ffdd060b2d69bd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141557
Tested-by: Jenkins
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>