tdf#151073 - enable firebird for appstore build w/o experimental mode
also don't default to hsqldb as embedded database when a build is done
without java support. Mainly to hide an UI bug - apparently the dialog
should show a non-supported message for known embedded drivers
at least that is what e13580dbff0e641e7807b81782a0a748e6d8e14e suggests
"
  // this indicates it's really a type which is known in general, but not supported on the current platform
  // show a message saying so
  //  eSpecialMessage = smUnsupportedType;
  insertDatasourceTypeEntryData( m_eCurrentSelection, sDisplayName );
  // remember this type so we can show the special message again if the user selects this
  // type again (without changing the data source)
  m_eNotSupportedKnownType = m_pCollection->determineType( m_eCurrentSelection );
"
but that doesn't work, at least not anymore. m_eNotSupportedKnownType was
since removed, no attempts are made in showing an error/warning message
returning the hsqldb driver results in an empty (no display string)
entry as the default element in the dropdown in the dialog, that still
allows to create a database file (but of course one that you cannot add
any actual database tables to) without any error message. At the very
least it is inconsistent with the rest of the dialog's code that clears
the input name for stuff that shouldn't be shown.

Change-Id: I8419888018be2a556c49d32e40f02c0ac1801930
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141631
Tested-by: Jenkins
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>