sw: make sure mark container is sorted before calling equal_range()
ooo89390-1.doc from crashtesting failed to export to ODT due to this.

The trouble is that equal_range() needs a sorted container, but a
listener may make the container unsorted.

Fix the problem by re-sorting the container after calling the listener.

A full ODT export is not needed to trigger the crash, it's enough to
hide deletions.

Change-Id: Id5601d7659d534c0bbfa84845abf0e6bb8e94c6f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141765
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins