tdf#83671 SmartArt: preserve pos and size for metafile
Import of a SmartArt creates a group with a background shape and then
the SmartArt shapes. The background shape determines the position and
size of the group.

The method Shape::convertSmartArtToMetafile() replaces the existing
SmartArt shapes with a metafile object and thereby throws away the
background shape. Thus the group takes it size from the metafile object,
which has default size 100x100 (Hmm) and position 0|0.
The patch remembers the original size and position from the background
shape and applies it then to the metafile shape.

Change-Id: Id7b1ccd20b2597c5e2adca7294a994fd2e0e066f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141830
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>