DOCX export, numbering portion format: consider full-para char formats as well
The bugdoc had a single paragraph with direct formatting (24pt font
size), numbering enabled and a bookmark covering half of the paragraph.
The numbering had default font size in the DOCX export result, not
inheriting from the paragraph.

In case the bookmark is removed, then the custom font size is not an
autoformat set on the whole text of the paragraph but rather a format on
the paragraph level, which does influence the formatting of the
numbering portion via the paragraph marker formatting, as expected.

Fix the problem in a way similar to what was done in commit
c69bafb18646a9e8e08afac124ccb2f941b96d69 (sw, numbering portion format:
consider full-para char formats as well, 2022-10-20), except that was
for the rendering and this is for the DOCX export.

This also required filtering out grab-bags in
lcl_writeParagraphMarkerProperties(), otherwise citation SDTs show up in
the export result at unexpected places, as shown by
CppunitTest_sw_ooxmlexport14's testTdf129353.

Change-Id: I7bb88e290f2a370d78ecc894f306bcb0a403545f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141995
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins