tdf#143377 Correct maximum Skew for TextOrientation in Calc
For zero or 180 degree text orinentation errors can happen
in the Border visualization, theoretically also in the
Text rendering. Ths has to do with sin(0) and sin(180)
being zero and lead internal to numerical problems, e.g.
a very huge Skew that when applied show the reported
'errors'.

I limit this mechanism now to +/- 1/2 degree from the
critical mentioned places, for Border and Text - to
not risk to have different points of corrections. The
UI only allows angles of 1 degree steps, but UNO API
and pdf import may allow more.

Change-Id: Idbc68f6a7beab84df0672165c2a813d96eeff84e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141999
Tested-by: Jenkins
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>