optimize text fitting algorithm to correctly calculate the fit
As we converted the chart stretching variable from int to double
this can cause the text fitting algorithm to calculate the fit
wrong. This commit changes the text fitting algorithm a bit so
that it produces similar result as before the change.

Change test testAutofittedTextboxIndent from 691200 to 712800 as
the fitting algorithm changed a bit. Visually the document still
looks correct.

Change-Id: Ib75733360039d80a5aed836f757c3f8e8ec56d61
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/142186
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>