Reinstate removal of lcl_GuessPaperSize
...originally done with 1d7964c5ff579ff93067dc1c3e45b522e3cb2b4e "CWS-TOOLING:
integrate CWS unifypaper01" on 2009-06-12.  However, the removed
lcl_GuessPaperSize then reappeared, apparently with the merge commit
4563270cfd37f1a08e0315a5624affa3a7680771 "CWS-TOOLING: integrate CWS
printerpullpages" on 2010-01-15, presumably due to a mis-merge somewhere along
the way.

SmModel::getRenderer and SmModel::render are e.g. called when doing "Export as
PDF..." of a Math Formula, but it seems a bit hard to get into the "no real
printer" branches there that would actually have called lcl_GuessPaperSize.  But
anyway, it looks reasonable to replace that function with the more elaborate
SvxPaperInfo::GetDefaultPaperSize.

(I stumbled over this when the code in lcl_GuessPaperSize caused a newly
introduced -Werror=dangling-reference with GCC 13 trunk, similar to
257e79b75189f1baff94935bfd8c719f0c1721d5 "-Werror=dangling-reference".  So I
figured I'd remove that code as apparently intended back in 2009 rather than try
to fix it now.)

Change-Id: Ia73d9ece8429bbeccd4f65a9efc156c0dee52c01
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/142246
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>