ofz#53016 Undefined-shift
before commit 44c6fc415616d0d7b72095615ac5f587354717fd
Date:   Wed Sep 21 20:54:00 2022 +0200

    split utl::TempFile into fast and named variants

    which makes it easier to know what each variant requires
    to stay on it's happy path

we closed the writable stream and got another readonly stream, after
that commit we have a read-write stream. With the "true" arg to
Storage it will Commit back to the underlying storage its contents
in its dtor. With a ReadOnly stream this failed and nothing happened,
with a read-write stream there is a write which didn't happen with a
read-only one.

Lets just use false here, we're not interested in keeping the contents.

"Direct" in this context is presumably analogous to the Compound
File Format "Direct mode":

"The mode in which files are accessed, either transacted or direct,
determines when changes are committed.

    Transacted mode uses a two-phase commit operation to make changes to
objects in a compound file, thereby keeping both the old and the new
copies of the document available until the user chooses to either save
or undo the changes.

    Direct mode incorporates changes to the document as they are made,
without the ability to later undo them.

For more information about access modes, see the OLE Programmer's Reference."

Change-Id: I1f710de227807073a6aff11a86d7cfaf6d380509
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/142283
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>