loplugin:stringconcatauto: There's no dangling-ref issue with O[U]StringNumber
...so it's unclear why 80fed61a318c4e4d44c4ba2f68a4282fc8ec35d2 "support
O(U)String::number() for fast string concatenation" added that here.

On the contrary, variables of type OStringNumber can be useful replacements for
calls to std::sprintf (which started to emit deprecation warnings with
macOS 13 SDK now), so this is in preparation for follow-up commits that will
replace many uses of that across the code base with various alternatives.

Change-Id: I6f8508d49dc84773c50f4c33dba38fe08c4c8969
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/142318
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>