tdf#125085 PPTX WordArt import: get shape font from run
PPTX uses the settings at the run for the text in a WordArt shape. LO
uses the settings of the shape. The patch copies the run font and
language properties to the shape.

In addition I have split the old method lcl_createPresetShape and
renamed the parts.

Now the method lcl_putCustomShapeIntoTextPathMode contains the geometry
changes to put a CustomShape into text path mode and to adapt the OOXML
values to the values needed for our preset Fontwork shapes, which are
based on binary MS Office.

MS Office has the style of the characters of a WordArt as properties
of the run. LibreOffice ignores most of the properties specified in
the style of a span element, when the text is displayed as Fontwork.
Instead LO uses the properties defined for the shape for styling the
text. Copying the text properties to the shape is now done in
method lcl_copyCharPropsToShape.

The values in testTdf125573_FontWorkScaleX depend on the used Font.
Since on creation time wrong Fonts were used and now correct Fonts
are imported, the values have to be updated.

Import of fill and stroke of the characters is still missing, only a
rudimentary import of color for solid fill exists.

Change-Id: I6f31fe07fb0656b2ce8581e7123265fa598ac9c1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/142421
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>