Fix VS2022 MSMs detection
After commits e12298684e138272428db12c3ca68054e47d78b9 (more VS2022 integration,
Oct 14 10:58:25 2021 +0200) and 2e98936b83f6d487d339b905d38f5f1994d6aa54 (Adapt
to latest VS 2022 MSM versions, Apr 08 09:15:19 2022 +0200), the new 143 version
of MSMs gets detected and passed to SCP, but not used there, so MSIs lack CRT.

Additionally, it may happen that a redist is configured that does not match the
actual CRT used in the build, because it picked a latest available version, and
then an error happens when building MSI, because it expects MSMs in wrong source
directory.

This simplifies the checks, makes sure to look for the correct MSM version, and
adds an error message in SCP in case when the passed version is not handled, to
detect the incomplete support early.

Change-Id: I47d45472309c584314730e7f5f46a9596fc8d21d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/142587
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>