SVG export: fix handling of semi-transparent, multi-para shape text
The bugdoc had a shape with two paragraphs. The export of the second paragraph
went wrong: the position & size of the text was larger than expected.

This problem was specific to semi-transparent text.
SVGTextWriter::setTextPosition() is recursive in this case. The non-transparent
case left nCurAction unchanged in the outer setTextPosition() and it returned 1
("text is found"). The semi-transparent case changed nCurAction by +2 and
returned 0 ("text is not found"). This led to all sorts of trouble, including
mismatching Push() and Pop() calls on the output device when replaying the
metafile.

Fix the problem by routing the empty state from the inner setTextPosition() to
the outer one: if we return 1 ("text is found"), then we know that the state is
non-empty. Once the empty state is correct in the outer setTextPosition(), then
even the transparent case also leaves nCurAction unchanged and the whole shape
text has the correct position and size.

I forgot to update this empty state in ther outer setTextPosition() in commit
019dd8b943b238ebcfcb13f0d4be845fc9b480b3 (SVG export: fix lost semi-transparent
text on shapes, 2020-07-17).

Change-Id: I7dc93fe13c4510220bf09a388e742799ed042510
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/142590
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins