gtktiledviewer: Fix getting tiles on Windows
The "CanvasSize + 1" hack was implemented by myself in commit
0a78002bb0eb98465fd77e16f50cda23f7bfb082 (Paint tiles on Windows;
Oct 24 09:01:32 2022 +0200) to workaround a rounding error, where
converting pixels to logic and back resulted in 1 pixel smaller
size (when I worked in Writer). This hack breaks in Calc, because
the size then is 1 pixel larger than expected, failing the assert
below.

This fixes the problem by disabling the map mode, as everywhere
else in such cases.

Change-Id: I37f0f03ae4cb19f2de0e7995064843d17b0c7189
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/142714
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>