WASM gbuild: write link dependencies into a file $@.linkdeps
The static build was added in commit
04c9982f539c4c83dd311dc7d10384b2e73c70a8 and commit
b8b0054759ddf939582c266bd1e8082f0736d52d and there are several files
written to workdir/LinkTarget/*/*.d.{libraries,externals,statics}
that list all the (recursive) dependencies of a LinkTarget.

Unfortunately the content of these files requires gbuild to interpret,
because it's using the Library names which may not map 1:1 to file
names, not to mention "externals"; this is all resolved by calls to
gb_LinkTarget_use_* anyway.

To get a command line that's usable outside of gbuild, write another
file $@.linkdeps as a side effect of gb_LinkTarget__command_dynamiclink.

Change-Id: I94f31142413561dcdb56fa722a4517f84fcd7d61
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/143393
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
Tested-by: Jenkins