svg export filter: when a text field is too small, its content is wrong
When the text field width is too small, the placeholder text spans
several lines.
That stopped the js engine to substitute the text field placeholder
with the right content.
This is a workaround but it should work in the majority of cases.
A complete solution needs to support svg text wrapping.

Change-Id: Ide52e08acd35b3764b8abc600e4b467acea0a248
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/142760
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Ashod Nakashian <ash@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/143050
Tested-by: Jenkins
Reviewed-by: Andras Timar <andras.timar@collabora.com>