Revert "TempFile: clear handle on close"
This reverts commit 7780251c7364fbdd48afa82f9b0b12762453e516:  As discussed in
the comments at
<https://gerrit.libreoffice.org/c/core/+/142344/2#message-c52fe7b2037e717525e89dd20a8f320142654ae0>
"TempFile: clear handle on close", that change was intended to reset *pHandle to
null in cases where osl_createTempFile returns an error and has already set
*pHandle.  But that is obviously not what the change actually does, it rather
sets *pHandle to null in (only some of the) cases where osl_createTempFile
returns an error and has not modified *pHandle.  Given that there may
potentially be clients that rely on osl_createTempFile not modifying the passed
in *pHandle when it returns an error, and given that the
sal/osl/unx/tempfile.cxx implementation of osl_createTempFile now inconsistently
set *pHandle to null in only some of the cases where it returns an error while
the sal/osl/w32/tempfile.cxx implementation of osl_createTempFile never sets
*pHandle to null when it returns an error, it appears best to just revert that
change.

Change-Id: Ia68353d04a488b024573ad03de0c5e5bd6e2a2c4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/142798
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>