tdf#151828 related: Make cut shape have the same name after paste
when possible

Fixes shape names without a number at the end of the name, having the
number 1 appended to the end of the name after cut and paste in the
same document.

When possible - because it might happen that a shape is inserted
after the cut operation and is given the same name as the cut shape
name. If so, when the shape is pasted into the same document it will
not have the same name as before it was cut.

Change-Id: I1d1ebc76d07d81da0286d1f0bb7d34ad973a43b8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/142983
Tested-by: Jenkins
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-by: Jim Raykowski <raykowj@gmail.com>